Farm manager/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
WTForms==3.1.2
Flask-WTF==1.2.1
requests==2.31.0 